Fixes setupResponseHeaders Method
During refactoring the httpMethod parameter was changed to HttpRequestBase but the variable not renamed; the request headers only were being iterated instead of both response/request headers after that; This removes the HttpRequestBase parameter, and instead iterates the response headers only.  Noticed this when doing a proxied download request during which the content-disposition header was being dropped.